Move expand_config and flatten_config to config module
