+++ vcspull/util.py
@@ -46,72 +46,6 @@ def in_dir(
     return configs
 
 
-def expand_config(config):
-    """Return expanded configuration.
-
-    end-user configuration permit inline configuration shortcuts, expand to
-    identical format for parsing.
-
-    :param config: the repo config in :py:class:`dict` format.
-    :type config: dict
-    :rtype: dict
-
-    """
-    for directory, repos in config.items():
-        for repo, repo_data in repos.items():
-
-            '''
-            repo_name: http://myrepo.com/repo.git
-
-            to
-
-            repo_name: { url: 'http://myrepo.com/repo.git' }
-
-            also assures the repo is a :py:class:`dict`.
-            '''
-
-            if isinstance(repo_data, string_types):
-                config[directory][repo] = {'url': repo_data}
-
-            '''
-            ``shell_command_after``: if str, turn to list.
-            '''
-            if 'shell_command_after' in repo_data:
-                if isinstance(repo_data['shell_command_after'], string_types):
-                    repo_data['shell_command_after'] = [
-                        repo_data['shell_command_after']
-                    ]
-
-    config = dict(
-        (os.path.expandvars(directory), repo_data) for
-        directory, repo_data in config.items()
-    )
-
-    config = dict(
-        (os.path.expanduser(directory), repo_data) for
-        directory, repo_data in config.items()
-    )
-
-    return config
-
-
-def flatten_config(config):
-    """Return config flattened into a list of dictionary.
-
-    :param config: expanded config
-    :type config: dict
-    :rtype: list of dict
-    """
-
-    repos = []
-    for repocwd, reponode in config.items():
-        for reponame, repo in reponode.items():
-            repo['name'] = reponame
-            repo['cwd'] = repocwd
-            repos.append(repo)
-    return repos
-
-
 def lookup_repos(config, dirmatch=None, vcsurlmatch=None, namematch=None):
     """Return a :py:obj:`list` list of repos from (expanded) config file.
